Creates custom hook to fetch pokemon data and creates Footer test.
