+++ tests/molecules/Footer.test.tsx
@@ -2,7 +2,15 @@ import { Footer } from "@/components/molecules/Footer";
 import { screen, render } from "tests/custom-render";
 
 describe("Footer component", () => {
-	it("Shows footer elements", () => {
+	it("Shows ", () => {
 		render(<Footer />);
+
+		const logos = screen.getAllByRole("img");
+		const gitHubLogo = logos[0];
+		const pokeApiLogo = logos[1];
+
+		expect(gitHubLogo).toBeInTheDocument();
+		expect(pokeApiLogo).toBeInTheDocument();
+		expect(pokeApiLogo).toHaveAttribute("alt", "PokeApi Logo");
 	});
 });
